requirements.txt
streamlit
requests
openai
numpy
pandas
plotly
gspread
google-auth
